added new MoveModel, MoveRepository and Move files to handle moving the oxen around the areas; applied auto format to all existing files; tweaked application.yml [changed hibernate dialect]; created a Makefile [temporarily until I figure out how gradle injects .env variables].

+++ src/main/java/com/jetbov/exercice/infra/dtos/UpdateOxDto.java
@@ -4,9 +4,9 @@ import com.jetbov.exercice.core.dtos.UpdateOx;
 
 public record UpdateOxDto(
         String earring,
-        double height
+        float height
 ) {
-    public UpdateOx toDomainRecord(){
+    public UpdateOx toDomainRecord() {
         return new UpdateOx(this.earring, this.height);
     }
 }
